pyproject.toml
[project]
name = "fms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "gunicorn>=23.0.0",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
    "whitenoise>=6.11.0",
    "django-crispy-forms>=2.3",
    "crispy-tailwind>=0.5",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "django-extensions>=4.1",
    "djlint>=1.36.4",
    "factory-boy>=3.3.3",
    "faker>=40.1.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]
